[FEAT] Enable daily rewards
