+++ src/features/game/components/modal/ModalProvider.tsx
@@ -18,7 +18,6 @@ import { DiscordBonus } from "features/game/expansion/components/DiscordBoat";
 import { Streams } from "./components/Streams";
 import { Merkl } from "./components/Merkl";
 import { Rewards } from "features/island/hud/components/referral/Rewards";
-import { DailyRewardChest } from "features/game/expansion/components/dailyReward/DailyReward";
 type GlobalModal =
   | "BUY_GEMS"
   | "DISCORD"
@@ -215,12 +214,6 @@ export const ModalProvider: FC<React.PropsWithChildren> = ({ children }) => {
       </Modal>
 
       <Rewards show={opened === "EARN"} onHide={handleClose} tab={"Earn"} />
-
-      <DailyRewardChest
-        show={opened === "DAILY_REWARD"}
-        onHide={handleClose}
-        tab={0}
-      />
     </ModalContext.Provider>
   );
 };
