+++ src/features/island/hud/components/referral/Rewards.tsx
@@ -1,35 +1,21 @@
 import { Modal } from "components/ui/Modal";
 import { CloseButtonPanel } from "features/game/components/CloseablePanel";
 import React, { useCallback, useContext, useState } from "react";
-// import giftIcon from "assets/icons/gift.png";
-import vipGift from "assets/decorations/vip_gift.png";
 import flowerIcon from "assets/icons/flower_token.webp";
 import loveCharmSmall from "assets/icons/love_charm_small.webp";
-import lockIcon from "assets/icons/lock.png";
-import { DailyRewardContent } from "../../../../game/expansion/components/dailyReward/DailyReward";
 import { SUNNYSIDE } from "assets/sunnyside";
 import { TaskBoard } from "./TaskBoard";
 import { MachineState } from "features/game/lib/gameMachine";
-import { rewardChestMachine } from "features/game/expansion/components/dailyReward/rewardChestMachine";
 
-import { Label } from "components/ui/Label";
-import { ButtonPanel } from "components/ui/Panel";
-import { getChapterTicket } from "features/game/types/chapters";
-import { VIPGiftContent } from "features/world/ui/VIPGift";
-import { BlockchainBox } from "./BlockchainBox";
-import { hasVipAccess } from "features/game/lib/vipAccess";
 import { pixelOrangeBorderStyle } from "features/game/lib/style";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import { useActor, useInterpret, useSelector } from "@xstate/react";
-import { getBumpkinLevel } from "features/game/lib/level";
+import { useSelector } from "@xstate/react";
 import { Context } from "features/game/GameProvider";
 import {
   IN_GAME_TASKS,
   InGameTaskName,
 } from "features/game/events/landExpansion/completeSocialTask";
 import { getKeys } from "features/game/types/decorations";
-import { useNow } from "lib/utils/hooks/useNow";
-import { hasFeatureAccess } from "lib/flags";
 import { DailyRewardClaim } from "features/game/components/DailyReward";
 
 const _chestCollectedAt = (state: MachineState) =>
@@ -97,12 +83,7 @@ export const Rewards: React.FC<Props> = ({ show, onHide, tab }) => {
         {currentTab === "Earn" && (
           <TaskBoard state={state} socialTasks={state.socialTasks} />
         )}
-        {currentTab === "Rewards" &&
-          (hasFeatureAccess(state, "DAILY_BOXES") ? (
-            <DailyRewardClaim />
-          ) : (
-            <RewardOptions />
-          ))}
+        {currentTab === "Rewards" && <DailyRewardClaim />}
         {/* {tab === 2 && <RewardsShop />} */}
       </CloseButtonPanel>
       {showMessage && (
@@ -138,149 +119,5 @@ export const Rewards: React.FC<Props> = ({ show, onHide, tab }) => {
         </div>
       )}
     </Modal>
-  );
-};
-
-export type RewardType = "DAILY_REWARD" | "VIP" | "BLOCKCHAIN_BOX";
-
-export const RewardOptions: React.FC<{ selectedButton?: RewardType }> = ({
-  selectedButton,
-}) => {
-  const { gameService } = useContext(Context);
-  const state = useSelector(gameService, (state) => state.context.state);
-  const now = useNow();
-  const chapterTicket = getChapterTicket(now);
-
-  // Get daily rewards data for chest state
-  const dailyRewards = useSelector(
-    gameService,
-    (state) => state.context.state.dailyRewards,
-  );
-  const bumpkinLevel = useSelector(gameService, (state) =>
-    getBumpkinLevel(state.context.state.bumpkin?.experience ?? 0),
-  );
-  const isRevealed = useSelector(gameService, (state) =>
-    state.matches("revealed"),
-  );
-
-  // Initialize chest service to check if chest is locked
-  const chestService = useInterpret(rewardChestMachine, {
-    context: {
-      // First code is 1
-      nextCode: dailyRewards?.chest?.code ?? 1,
-      openedAt: dailyRewards?.chest?.collectedAt ?? 0,
-      bumpkinLevel,
-    },
-  });
-  const [chestState] = useActor(chestService);
-
-  const [selected, setSelected] = useState<RewardType | undefined>(
-    selectedButton,
-  );
-  const { t } = useAppTranslation();
-
-  if (selected === "DAILY_REWARD") {
-    return (
-      <DailyRewardContent
-        onClose={() => {
-          setSelected(undefined);
-        }}
-        gameService={gameService}
-        dailyRewards={state.dailyRewards}
-        isRevealed={isRevealed}
-        bumpkinLevel={bumpkinLevel}
-        chestService={chestService}
-        chestState={chestState}
-      />
-    );
-  }
-
-  if (selected === "VIP") {
-    return <VIPGiftContent onClose={() => setSelected(undefined)} />;
-  }
-
-  if (selected === "BLOCKCHAIN_BOX") {
-    return <BlockchainBox />;
-  }
-
-  const vipOpenedAt = state.pumpkinPlaza.vipChest?.openedAt ?? 0;
-
-  const hasOpenedVIP =
-    !!vipOpenedAt &&
-    new Date(vipOpenedAt).toISOString().substring(0, 7) ===
-      new Date().toISOString().substring(0, 7);
-
-  const today = new Date().toISOString().substring(0, 10);
-  const hasOpenedDaily =
-    new Date(state.dailyRewards?.chest?.collectedAt ?? 0)
-      .toISOString()
-      .substring(0, 10) === today;
-
-  return (
-    <>
-      <Label type="default" className="mb-1">
-        {t("rewards.claim.title")}
-      </Label>
-
-      <ButtonPanel
-        onClick={
-          bumpkinLevel < 6 ? undefined : () => setSelected("DAILY_REWARD")
-        }
-        disabled={bumpkinLevel < 6}
-        className="mb-1"
-      >
-        <div className="flex items-start">
-          <img
-            src={SUNNYSIDE.decorations.treasure_chest}
-            className="w-10 mr-4"
-          />
-          <div className="relative flex-1">
-            <p className="text-sm mb-1">{t("rewards.daily.title")}</p>
-            <p className="text-xs">
-              {t("rewards.daily.description", { ticket: chapterTicket })}
-            </p>
-          </div>
-          {hasOpenedDaily && (
-            <Label className="absolute top-0 right-0" type="success">
-              {t("rewards.daily.claimed")}
-            </Label>
-          )}
-          {bumpkinLevel < 6 && (
-            <Label
-              icon={lockIcon}
-              secondaryIcon={SUNNYSIDE.icons.player}
-              className="absolute top-0 right-1"
-              type="formula"
-            >
-              {`${t("lvl")} 6`}
-            </Label>
-          )}
-        </div>
-      </ButtonPanel>
-
-      <ButtonPanel onClick={() => setSelected("VIP")}>
-        <div className="flex items-start">
-          <img src={vipGift} className="w-10 mr-4" />
-          <div className="relative flex-1">
-            <p className="text-sm mb-1">{t("rewards.vip.title")}</p>
-            <p className="text-xs">{t("rewards.vip.description")}</p>
-          </div>
-          {hasOpenedVIP && (
-            <Label className="absolute top-0 right-0" type="success">
-              {t("rewards.vip.claimed")}
-            </Label>
-          )}
-          {!hasOpenedVIP && !hasVipAccess({ game: state, now }) && (
-            <Label
-              icon={lockIcon}
-              className="absolute top-0 right-0"
-              type="formula"
-            >
-              {t("locked")}
-            </Label>
-          )}
-        </div>
-      </ButtonPanel>
-    </>
   );
 };
